pyproject.toml
[tool.poetry]
name = "chzzk-streaming-manager"
version = "0.1.0"
description = "chzzk chating manager by AI Agent"
authors = ["kr-MATAGI <dev_matagi@kakao.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.5.1"
langchain = "^0.3.14"
langgraph = "^0.2.64"
python-dotenv = "^1.0.1"
langchain-openai = "^0.3.0"
ipython = "^8.31.0"
graphviz = "^0.20.3"
pillow = "^11.1.0"
matplotlib = "^3.10.0"
selenium = "^4.27.1"
bs4 = "^0.0.2"
webdriver-manager = "^4.0.2"
psycopg2-binary = "^2.9.10"
aioredis = "^2.0.1"
redis = "^5.2.1"
asyncpg = "^0.30.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
